+++ addons/website_forum/static/src/js/website_forum.editor.js
@@ -2,8 +2,11 @@ odoo.define('website_forum.editor', function (require) {
 "use strict";
 
 var core = require('web.core');
-var wUtils = require('website.utils');
+var Widget = require('web.Widget');
+var SummernoteManager = require('web_editor.rte.summernote');
 var WebsiteNewMenu = require("website.newMenu");
+var wUtils = require('website.utils');
+var websiteRootData = require('website.WebsiteRoot');
 
 var _t = core._t;
 
@@ -54,4 +57,15 @@ WebsiteNewMenu.include({
         });
     },
 });
+
+var WebsiteForumManager = Widget.extend({
+    /**
+     * @override
+     */
+    init: function () {
+        this._super.apply(this, arguments);
+        new SummernoteManager(this);
+    },
+});
+websiteRootData.websiteRootRegistry.add(WebsiteForumManager, '#wrapwrap');
 });
